[refactor] Re-organized ShortAbout JSX.
