+++ src/components/Home/index.tsx
@@ -1,4 +1,4 @@
-import { ShortAbout } from './ShortAbout';
+import { ShortAbout } from '../ShortAbout/ShortAbout';
 import { WorkPreview } from './WorkPreview';
 
 export { ShortAbout, WorkPreview };
